chore: move sql-editor setting button to the top-right corner
