+++ frontend/src/views/sql-editor/AsidePanel/GutterBar/GutterBar.vue
@@ -15,27 +15,17 @@
 
     <div class="flex flex-col justify-end items-center">
       <OpenAIButton :size="size" />
-
-      <SettingButton
-        v-if="!hideSettingButton"
-        :style="buttonStyle"
-        v-bind="buttonProps"
-      />
     </div>
   </div>
 </template>
 
 <script setup lang="ts">
-import { storeToRefs } from "pinia";
-import { computed, toRef } from "vue";
-import { useAppFeature, useSQLEditorStore } from "@/store";
-import { SettingButton } from "../../Setting";
 import { useSQLEditorContext, type AsidePanelTab } from "../../context";
 import OpenAIButton from "./OpenAIButton.vue";
 import TabItem from "./TabItem.vue";
-import { useButton, type Size } from "./common";
+import { type Size } from "./common";
 
-const props = withDefaults(
+withDefaults(
   defineProps<{
     size?: Size;
   }>(),
@@ -45,25 +35,6 @@ const props = withDefaults(
 );
 
 const { asidePanelTab } = useSQLEditorContext();
-const { strictProject } = storeToRefs(useSQLEditorStore());
-const disableSetting = useAppFeature("bb.feature.sql-editor.disable-setting");
-
-const { props: buttonProps, style: buttonStyle } = useButton({
-  size: toRef(props, "size"),
-  active: false,
-  disabled: false,
-});
-
-const hideSettingButton = computed(() => {
-  if (disableSetting.value) {
-    return true;
-  }
-  if (strictProject.value) {
-    return true;
-  }
-
-  return false;
-});
 
 const handleClickTab = (target: AsidePanelTab) => {
   asidePanelTab.value = target;
