+++ frontend/src/views/sql-editor/Setting/SettingButton.vue
@@ -1,18 +1,16 @@
 <template>
   <NButton v-if="show" @click="goSetting">
-    <template #icon>
-      <SettingsIcon />
-    </template>
+    <SettingsIcon class="w-4 h-4" />
   </NButton>
 </template>
 
 <script setup lang="ts">
 import { SettingsIcon } from "lucide-vue-next";
+import { NButton } from "naive-ui";
 import { computed } from "vue";
 import { useRouter } from "vue-router";
 import { SQL_EDITOR_SETTING_MODULE } from "@/router/sqlEditor";
 import { useSidebarItems } from "./Sidebar";
-import { NButton } from "naive-ui";
 
 const router = useRouter();
 const { itemList } = useSidebarItems();
